Extract nested classes from CachingHiveMetastore
